refactor: move readGitHubCliToken to copilot/auth.ts
GitHub-specific token reading logic now lives alongside other
Copilot auth helpers instead of in general config layer.
